+++ moor/test/data/utils/expect_generated.dart
@@ -2,10 +2,14 @@ import 'package:moor/moor.dart';
 import 'package:test/test.dart';
 
 extension ComponentExpectations on Component {
-  void expectGenerates(String sql) {
+  void expectGenerates(String sql, [List<dynamic> variables]) {
     final ctx = GenerationContext(SqlTypeSystem.defaultInstance, null);
     writeInto(ctx);
 
     expect(ctx.sql, sql);
+
+    if (variables != null) {
+      expect(ctx.boundVariables, variables);
+    }
   }
 }